refactor: auth 패키지 컨벤션 맞추어 통일
