+++ src/main/java/com/bookbla/americano/domain/auth/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.bookbla.americano.domain.auth.controller;
 
 import com.bookbla.americano.domain.auth.controller.dto.request.LoginRequestDto;
 import com.bookbla.americano.domain.auth.controller.dto.response.LoginResponseDto;
-import com.bookbla.americano.domain.auth.service.AuthService;
+import com.bookbla.americano.domain.auth.service.impl.AuthServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/auth")
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthServiceImpl authService;
 
     @PostMapping("/login/{oAuthType}")
     public ResponseEntity<LoginResponseDto> login(
