+++ src/main/java/com/bookbla/americano/domain/auth/infra/kakao/KakaoOAuth2Provider.java
@@ -5,8 +5,8 @@ import com.bookbla.americano.base.exception.BaseException;
 import com.bookbla.americano.base.exception.BaseExceptionType;
 import com.bookbla.americano.domain.auth.infra.kakao.dto.KakaoOAuth2MemberResponse;
 import com.bookbla.americano.domain.auth.service.OAuth2Provider;
-import com.bookbla.americano.domain.auth.service.dto.KakaoMemberResponse;
-import com.bookbla.americano.domain.auth.service.dto.KakaoTokenResponse;
+import com.bookbla.americano.domain.auth.infra.kakao.dto.KakaoMemberResponse;
+import com.bookbla.americano.domain.auth.infra.kakao.dto.KakaoTokenResponse;
 import com.bookbla.americano.domain.auth.service.dto.OAuth2MemberResponse;
 import com.bookbla.americano.domain.member.enums.MemberType;
 import java.util.Objects;
