+++ src/main/java/com/bookbla/americano/domain/auth/infra/kakao/dto/KakaoMemberResponse.java
@@ -1,4 +1,4 @@
-package com.bookbla.americano.domain.auth.service.dto;
+package com.bookbla.americano.domain.auth.infra.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
